test: split Node tests into multiple files, migrate to AVA
