+++ packages/zwave-js/src/lib/test/assertCC.ts
@@ -1,4 +1,5 @@
 import type { CCConstructor, CommandClass } from "@zwave-js/cc";
+import type { ExecutionContext } from "ava";
 import { SendDataBridgeRequest } from "../serialapi/transport/SendDataBridgeMessages";
 import { SendDataRequest } from "../serialapi/transport/SendDataMessages";
 
@@ -26,6 +27,34 @@ export function assertCC<
 	if (options.ccValues) {
 		for (const [prop, val] of Object.entries(options.ccValues)) {
 			expect((command as any)[prop]).toBe(val);
+		}
+	}
+}
+
+export function assertCCAva<
+	TConst extends CCConstructor<CommandClass> = CCConstructor<CommandClass>,
+>(
+	t: ExecutionContext,
+	callArg: any,
+	options: {
+		nodeId?: number;
+		cc: TConst;
+		ccValues?: Record<string, any>;
+	},
+): void {
+	const request: SendDataRequest | SendDataBridgeRequest = callArg;
+	try {
+		t.true(request instanceof SendDataRequest);
+	} catch {
+		t.true(request instanceof SendDataBridgeRequest);
+	}
+	if (options.nodeId) t.is(request.getNodeId(), options.nodeId);
+
+	const command = request.command;
+	t.true(command instanceof options.cc);
+	if (options.ccValues) {
+		for (const [prop, val] of Object.entries(options.ccValues)) {
+			t.is((command as any)[prop], val);
 		}
 	}
 }
